+++ ICEngine/src/Core/Data.h
@@ -6,10 +6,10 @@
 void* ICE_Data_Insert(ICE_State * state, size_t _size);
 
 /// Return pointer to a Data
-void* ICE_Data_Get(ICE_State * state, unsigned int nb_data);
+void* ICE_Data_Get(ICE_State * state, size_t nb_data);
 
 /// Destroy a Data
-void ICE_Data_Destroy(ICE_State * state, unsigned int nb_data);
+void ICE_Data_Destroy(ICE_State * state, size_t nb_data);
 
 void ICE_Data_DestroyAll(ICE_State * state);
